+++ essentials-framework/Essentials Devices Common/Essentials Devices Common/VideoCodec/CiscoCodec/CiscoSparkCodec.cs
@@ -2470,20 +2470,6 @@ ConnectorID: {2}"
     }
 
 
-	/// <summary>
-    /// Represents a codec command that might need to have a friendly label applied for UI feedback purposes
-    /// </summary>
-    public class CodecCommandWithLabel
-    {
-        public string Command { get; set; }
-        public string Label { get; set; }
-
-        public CodecCommandWithLabel(string command, string label)
-        {
-            Command = command;
-            Label = label;
-        }
-    }
 
     /// <summary>
     /// Tracks the initial sycnronization state of the codec when making a connection
