+++ essentials-framework/Essentials Devices Common/Essentials Devices Common/VideoCodec/Interfaces/IHasSelfviewSize.cs
@@ -1,4 +1,4 @@
-﻿using PepperDash.Essentials.Devices.Common.VideoCodec.Cisco;
+﻿using PepperDash.Essentials.Devices.Common.VideoCodec;
 
 namespace PepperDash.Essentials.Core.DeviceTypeInterfaces
 {
