+++ essentials-framework/Essentials Devices Common/Essentials Devices Common/VideoCodec/VideoCodecBase.cs
@@ -1995,4 +1995,19 @@ namespace PepperDash.Essentials.Devices.Common.VideoCodec
 			}
 		}
 	}
+    /// <summary>
+    /// Represents a codec command that might need to have a friendly label applied for UI feedback purposes
+    /// </summary>
+    public class CodecCommandWithLabel
+    {
+        public string Command { get; set; }
+        public string Label { get; set; }
+
+        public CodecCommandWithLabel(string command, string label)
+        {
+            Command = command;
+            Label = label;
+        }
+    }
+
 }
